+++ tests/testthat/test-dependencies.R
@@ -159,13 +159,13 @@ test_that("Suggests are dev. deps for non-package projects", {
   expect_true(deps$Dev)
 })
 
-test_that("Suggests are _not_ dev. deps for package projects", {
+test_that("Suggests are dev. deps for package projects", {
   renv_tests_scope()
   writeLines(c("Type: Package", "Suggests: bread"), con = "DESCRIPTION")
-  deps <- dependencies(dev = FALSE)
+  deps <- dependencies(dev = TRUE)
   expect_true(nrow(deps) == 1)
   expect_true(deps$Package == "bread")
-  expect_false(deps$Dev)
+  expect_true(deps$Dev)
 })
 
 test_that("packages referenced by modules::import() are discovered", {